requirements.txt
numpy==1.25.2
pandas==2.1.0
seaborn==0.12.2
matplotlib==3.7.2
joblib==1.3.2
boto3==1.28.48
scikit-learn==1.2.0
Flask==2.3.3
python-dotenv==1.0.0
packaging==23.1
pyaes==1.6.1
python-dateutil==2.8.2